RVM-2947 add message shape interfaces for each rvm publish call
This is a first step to rein in the complexity around interfacing with the rvm protocol.
Bascially here we just type out the shape of the current api calls. The next step would
involve reviewing the patterns here and forming an agreed upon format that we share with
RVM team as well as consolidating the interal usage to a single interface. This also
contains a JS-TS conversions of app_assets