RUN-3014 deprecate rvm send  (#28)
* RUN-2991 move the rvm message bus file to typescript

This also includes all the refactor required to the import /
require statements.

* RUN-2991 update utils import

* RUN-2991 lint

* RUN-3006 send shapes added

* RUN-3006 switching to fix another PR, brb

* RUN-3006 add message shape interfaces for each rvm publish call

This is a first step to rein in the complexity around interfacing with the rvm protocol.
Bascially here we just type out the shape of the current api calls. The next step would
involve reviewing the patterns here and forming an agreed upon format that we share with
RVM team as well as consolidating the interal usage to a single interface. This also
contains a JS-TS conversions of app_assets

* RUN-3006 linting...

* RUN-3006 linting, add to license list

* RUN-3006 linting

* RUN-3014 switch over all instances of rvm bus.send to .publish