RUN-4400: Call Unix binaries with full path
It turns out that within the Node adapter test framework, there's a bunch of
layers that result losing track of the PATH environment variable.

Instead of trying to find out where, we could simply ignore the non-existente
PATH and call the binaries that we know are there.

This assumes a Unix system that's not _too_ far out of whack.