RUN-4538 explicit close of the download window on completion
A comment from the code explaines it:
Windows that are created from a window.open that trigger a file
download have their url set to "" here no matter what the requested
url was. This is not true if the file download was triggered via setting
location.href to the the desired file url. In this case the url will
be that of page it self, NOT the download url.

There is an extra process created for the download in the href case as
well that does not go away immediately (this is true in both electron
and chrome). We just need to be sure that we dont close the parent
window in the href case as that is what the browser window will map to
if the url is present.

There may be some value in diving in a bit deeper on the native side to
see if there is a spot to kill it from there without having tt reach
around and do it from the core.