RVM-733: User-configurable flush interval for app logs
If `startup_app.appLogFlushInterval` is defined, it'll be used as a flush
interval for application logs, with the value being interpreted as
milliseconds. Note that flushing can still occur earlier if a buffer of 1
MB is hit (maybe we want to also make that magic value configurable?).

If it's not defined, the existing flush interval of ten seconds is used.