RUN-4596 RUN-4742 Flag-protected cache migration (#631)
* RUN-4596, RUN-4742: Flag-protected cache migration

This won't do anything unless both old-local-storage-path and
new-local-storage-path are specified, in which case either the RVM
determined cache migration should be done by the runtime, or the user very
deliberately wanted the runtime to do cache migration.

See https://github.com/openfin/runtime/pull/1250 for gory details.

* RVM-733: User-configurable flush interval for app logs

If `startup_app.appLogFlushInterval` is defined, it'll be used as a flush
interval for application logs, with the value being interpreted as
milliseconds. Note that flushing can still occur earlier if a buffer of 1
MB is hit (maybe we want to also make that magic value configurable?).

If it's not defined, the existing flush interval of ten seconds is used.