RUN-4765 Don't use decorated fs module; use the original
Electron decorate's NodeJS's fs module to handle asars. Since we are
adding on functionality to the fs module to also verify asars, this
causes an issue where for every file that needs to be read in via
fs.readFileSync, there needs to be a matching file that contains a
hashed signature. This prevents us from using the decorated fs module
for general purpose file reading.

Instead, we expose 'original-fs' in the case you don't need to read
files from asars.