RUN-5363: Enable Electron's unit tests (#839)
Goal of this task is to add possibility to launch Electron's unit tests located in directory electron/spec on OpenFin's application. It will allow us to check if our changes break their apis or cause any regressions. These tests - in the future - could also be used as integration criteria for our tasks/bugfixes.